creating an additional page in navbar for Edaucation, will later be moved into CV, doesnt display edaucationList for the moment
