+++ src/ProjectCard.jsx
@@ -2,7 +2,7 @@ import React from "react";
 
 const ProjectCard = ({ project }) => {
   return (
-   <>
+    <>
       <div class="ui card">
         <div class="image">
           <img src={project.image} />
@@ -10,14 +10,18 @@ const ProjectCard = ({ project }) => {
         <div class="content">
           <h3 class="ui header">{project.name}</h3>
           <div class="description">{project.description}</div>
-          
-        <a href={project.git} target="_blank"><button type="button" id="github">GitHub</button></a>
-        <a href={project.web} target="_blank"><button type="button" id="webpage">WebPage</button></a>
-        </div>
-      
-        
-        
 
+          <a href={project.git} target="_blank">
+            <button type="button" id="github">
+              GitHub
+            </button>
+          </a>
+          <a href={project.web} target="_blank">
+            <button type="button" id="webpage">
+              WebPage
+            </button>
+          </a>
+        </div>
       </div>
     </>
   );
