+++ src/data/eduacations.json
@@ -1,0 +1,25 @@
+[
+  {
+    "id": 1,
+    "name": "Craft Academy",
+    "description": "Java Script, Ruby",
+    "time": "2020",
+    "title": "Full stack Developer"
+  },
+
+  {
+    "id": 2,
+    "name": "Eleiko",
+    "description": "Personal Trainer, Strenght Coach, Powerlifting",
+    "time": "2015",
+    "title": "PT"
+  },
+
+  {
+    "id": 3,
+    "name": "Stockholm University, geochemical institution",
+    "description": "geo geek",
+    "time": "2012",
+    "title": "Quaternary geology and climate development,"
+  }
+]
