+++ src/index.js
@@ -6,7 +6,7 @@ import Footer from "./Footer";
 import Projects from "./Projects";
 import Cv from "./Cv";
 import Contact from "./Contact";
-import Experience from "./Experience";
+import Eduacations from "./Eduacations";
 import { Switch, Route, BrowserRouter } from "react-router-dom";
 
 const App = () => {
@@ -18,9 +18,8 @@ const App = () => {
         <Route exact path="/cv" component={Cv}></Route>
         <Route exact path="/projects" component={Projects}></Route>
         <Route exact path="/contact" component={Contact}></Route>
-        <Route exact path="/experience" component={Experience}></Route>
+        <Route exact path="/eduacations" component={Eduacations}></Route>
       </Switch>
-
       <Footer />
     </>
   );
